+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 0.8.0 - 2019-08-07
 ### Added
 - Add `llhd-check` tool to verify consistency of assembly files.
 - Add `const iN$` flavor to generate a constant integer signal.